+++ Changelog.md
@@ -11,6 +11,7 @@ OpenCore Changelog
 - Added AppleEvent mouse support in OpenCanopy
 - Fixed AppleEvent and OpenCanopy compatibility with OVMF TPL restrictions
 - Added mouse drivers to the package as OVMF needs one
+- Added memory region reservation support
 
 #### v0.5.7
 - Added TimeMachine detection to picker